requirements_fixed.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sendgrid==6.10.0
Authlib==1.2.1
dnspython==2.4.2
requests==2.31.0
aiohttp>=3.9.2
pydantic[email]==2.5.0
gunicorn==21.2.0
python-multipart>=0.0.6
jinja2>=3.1.2